requirements.txt
Flask>=3.0,<4.0
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
